Insightful Pulse: An Blogging Application
I have created a Insightful Pulse which is an application where users can create their account using OAUTH provided by Google and manually using their email. This application helps them to create thinking and experience to their blogs so that a reader can read it to gain some knowledge from them. Also they can like, share and comment on the blogs. In this application, I have used a Tech Stack like HTML, CSS, JAVASCRIPT and REACT.js and firebase Realtime Database and its storage for storing images and files and used Firebase Authentication to manage user access. This is not a role based application. It is simple, efficient application for only blogs and using the datasets of the blogs, we have also created a home page of blogs by analyzing latest blogs and images and users on this website which helps a reader to read latest blogs and view latest images which built a confidence in them to use this application as much as required.